+++ common/src/main/kotlin/com/cobblemon/mod/common/CobblemonItems.kt
@@ -20,13 +20,21 @@ import com.cobblemon.mod.common.pokemon.helditem.CobblemonHeldItemManager
 import com.cobblemon.mod.common.registry.CompletableRegistry
 import dev.architectury.registry.registries.RegistrySupplier
 import net.minecraft.block.Block
-import net.minecraft.item.AliasedBlockItem
-import net.minecraft.item.BlockItem
-import net.minecraft.item.Item
-import net.minecraft.item.ItemGroup
+import net.minecraft.block.ComposterBlock
+import net.minecraft.item.*
 import net.minecraft.util.registry.Registry
 
 object CobblemonItems : CompletableRegistry<Item>(Registry.ITEM_KEY) {
+
+    init {
+        this.completed.thenRun {
+            val compostChance = ComposterBlock.ITEM_TO_LEVEL_INCREASE_CHANCE.getFloat(Items.OAK_LEAVES)
+            // Should always pass unless Mojang reworks leaves to no longer work in the Composter, in that case we already updated w.o doing anything
+            if (compostChance != ComposterBlock.ITEM_TO_LEVEL_INCREASE_CHANCE.defaultReturnValue()) {
+                ComposterBlock.ITEM_TO_LEVEL_INCREASE_CHANCE[APRICORN_LEAVES.get()] = compostChance
+            }
+        }
+    }
 
     private val pokeballs = mutableListOf<RegistrySupplier<PokeBallItem>>()